EmpleabotPlus-main/requirements.txt
azure-servicemanagement-legacy==0.20.8
email-validator==2.2.0
flask==3.1.0
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
openai==1.66.3
psycopg2-binary==2.9.10
PyPDF2==3.0.1
trafilatura==2.0.0
twilio==9.5.0
werkzeug==3.1.3 
python-dotenv==1.0.1